refactor(server): Move JsonPointerLocator to zally-core #560
