+++ server/src/main/java/de/zalando/zally/rule/RulesValidator.kt
@@ -8,6 +8,7 @@ import de.zalando.zally.core.ContentParseResult.ParsedSuccessfully
 import de.zalando.zally.core.ContentParseResult.ParsedWithErrors
 import de.zalando.zally.rule.api.Severity
 import de.zalando.zally.rule.api.Violation
+import de.zalando.zally.core.JsonPointerLocator
 import org.slf4j.LoggerFactory
 import java.lang.reflect.InvocationTargetException
 import java.net.URI
